requirements.txt
pynacl
requests
